Move files unused by OCM from /common to /cim
