+++ src/cim/components/ClusterDeployment/index.ts
@@ -3,6 +3,7 @@ export * from './helpers';
 export * from './AdditionalNTPSourcesDialogToggle';
 export * from './LogsDownloadButton';
 export * from './wizardTransition';
+export * from './networkConfigurationValidation';
 
 export { default as ClusterDeploymentValidationsOverview } from './ClusterDeploymentValidationsOverview';
 export { default as ClusterDeploymentDetails } from './ClusterDeploymentDetails';
@@ -11,6 +12,7 @@ export { default as ClusterDeploymentCredentials } from './ClusterDeploymentCred
 export { default as ClusterDeploymentKubeconfigDownload } from './ClusterDeploymentKubeconfigDownload';
 export { default as ClusterDeploymentWizard } from './ClusterDeploymentWizard';
 export { default as ClusterInstallationError } from './ClusterInstallationError';
+export { default as NetworkConfiguration } from './NetworkConfiguration';
 
 export { default as ACMClusterDeploymentDetailsStep } from './ACMClusterDeploymentDetailsStep';
