+++ src/common/components/clusterConfiguration/index.ts
@@ -1,8 +1,6 @@
 export { default as SNOControlGroup } from './SNOControlGroup';
 export { default as SNODisclaimer } from './SNODisclaimer';
 export { default as OpenShiftVersionSelect } from './OpenShiftVersionSelect';
-export { default as NetworkConfigurationFormFields } from './NetworkConfigurationFormFields';
-export { default as NetworkConfiguration } from './NetworkConfiguration';
 export { default as UploadSSH } from './UploadSSH';
 export { default as ProxyFields, ProxyInputFields } from './ProxyFields';
 export { default as DownloadIsoModal } from './DownloadIsoModal';
@@ -12,6 +10,5 @@ export { default as FormatDiskWarning } from './FormatDiskWarning';
 export { default as SecurityFields } from './SecurityFields';
 
 export * from './utils';
-export * from './networkConfigurationValidation';
 export * from './DiscoveryTroubleshootingModal';
 export * from './DiscoveryImageConfigForm';
